requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
billiard==3.6.4.0
    # via celery
celery==5.1.2
    # via -r requirements/requirements.in
certifi==2021.5.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.4
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==3.2.7
    # via
    #   -r requirements/requirements.in
    #   django-cors-headers
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==3.8.0
    # via -r requirements/requirements.in
django-redis==5.0.0
    # via -r requirements/requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements/requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==4.8.0
    # via -r requirements/requirements.in
drf-yasg==1.20.0
    # via -r requirements/requirements.in
feedparser==6.0.8
    # via -r requirements/requirements.in
gevent==21.8.0
    # via -r requirements/requirements.in
greenlet==1.1.1
    # via gevent
gunicorn==20.1.0
    # via -r requirements/requirements.in
idna==3.2
    # via requests
inflection==0.5.1
    # via drf-yasg
iniconfig==1.1.1
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.0.1
    # via coreschema
kombu==5.1.0
    # via celery
markupsafe==2.0.1
    # via jinja2
packaging==21.0
    # via
    #   drf-yasg
    #   pytest
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.20
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements/requirements.in
py==1.10.0
    # via pytest
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements/requirements.in
    #   pytest-django
pytest-django==4.4.0
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via -r requirements/requirements.in
python-dotenv==0.19.0
    # via -r requirements/requirements.in
pytz==2021.1
    # via
    #   celery
    #   django
redis==3.5.3
    # via django-redis
requests==2.26.0
    # via coreapi
ruamel.yaml==0.17.16
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sentry-sdk==1.3.1
    # via -r requirements/requirements.in
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pytest
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.6
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
zope.event==4.5.0
    # via gevent
zope.interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
